CheckersMoveLauncher implemented
The CheckersMoveLauncher is used like a function that stores state
until it is ready to execute its callback. It has a single public
function that will set the source move, then the destination move,
and then call its callback.

Upon calling its callback, the launcher is then reset and ready
to accept another move set.